fix: windows comparison in json
