+++ tests/boltz_output.py
@@ -1,5 +1,6 @@
 from pathlib import Path
-from tests.utils import normalize_text_paths, load_yaml, normalize_msa_paths
+from tests.utils import normalize_text_paths, load_yaml
+from typing import Any, Dict
 
 REPO_ROOT = Path(__file__).resolve().parent.parent
 OUTPUT_DIR = REPO_ROOT / "notebooks/boltz_test"
@@ -12,6 +13,21 @@ reference_sh = REPO_ROOT / "references/boltz/run.sh"
 
 
 MARKER = "boltz_test"
+
+
+def normalize_msa_paths(yaml_dict: Dict[str, Any], marker: str) -> Dict[str, Any]:
+    """
+    Strip absolute prefixes from msa fields only.
+    This is very tailored to the structure of the input YAML produced in the Boltz notebook, but it is sufficient for our testing purposes.
+    """
+    for seq in yaml_dict.get("sequences", []):
+        protein = seq.get("protein", {})
+        msa = protein.get("msa")
+        if msa and marker in msa:
+            prefix, suffix = msa.split(marker, 1)
+            # replace backslashes in the suffix only to please Windows
+            protein["msa"] = marker + suffix.replace("\\", "/")
+    return yaml_dict
 
 
 def test_run_sh_exists():
